Set fail-fast to false for GHA jobs
In the `test` and `build` jobs of the `build` workflow we leverage the strategy
matrix feature to parallelize runs on different Python versions. However, the
default value of `true` for the `fail-fast` option means that if one of the
runs fails then all others will be canceled. This changes that behavior so all
runs will be allowed to finish even if some runs fail.